Add a simple test for peer discovery algorithm
Simply copy a node_bl_test_gossip.cpp to node_bl_test_peerconnection.cpp,
and remove gossip feature from it.
Also, append a build script in CMakeLists.txt.
Fix a server.js which is failed to initialize a visualizationData for small sized experiment.
This commit is to make a demo video for BLEEP 1-year evaluation.